requirements.txt
black==18.9b0
pyenet==1.3.13.post6
twisted==18.9.0
